# lib/api-zod/src/generated/types/attributionEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketing OS API - Attribution Engine & Leads Dashboard
 * OpenAPI spec version: 0.1.0
 */
import type { AttributionEventEventType } from "./attributionEventEventType";
import type { AttributionEventMatchLevel } from "./attributionEventMatchLevel";

export interface AttributionEvent {
  id: number;
  tenantId: number;
  eventType: AttributionEventEventType;
  gclid?: string | null;
  wbraid?: string | null;
  fbclid?: string | null;
  hashedPhone?: string | null;
  hashedEmail?: string | null;
  billingAddress?: string | null;
  utmSource?: string | null;
  utmCampaign?: string | null;
  utmMedium?: string | null;
  landingPage?: string | null;
  matchLevel?: AttributionEventMatchLevel;
  matchConfidence?: number | null;
  createdAt: Date;
}