# lib/api-zod/src/generated/types/createTenantInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketing OS API - Attribution Engine & Leads Dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTenantInput {
  name: string;
  serviceTitanId?: string;
  timezone?: string;
}