# lib/api-zod/src/generated/types/tenant.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Marketing OS API - Attribution Engine & Leads Dashboard
 * OpenAPI spec version: 0.1.0
 */

export interface Tenant {
  id: number;
  name: string;
  serviceTitanId?: string | null;
  timezone: string;
  isActive: boolean;
  createdAt: Date;
}